requirements.txt
astroid==2.4.2
blinker==1.4
Click==7.0
colorama==0.4.3
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-SQLAlchemy>=2.3.2
Flask-WTF==0.14.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.4.3
MarkupSafe==1.1.0
mccabe==0.6.1
pylint==2.5.3
python-dotenv==0.13.0
six==1.15.0
SQLAlchemy==1.2.15
toml==0.10.1
typed-ast==1.4.1
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.12.1
WTForms==2.2.1
